Remove NestUI
- s.t.p.nest.Colors becomes s.t.p.ConsoleLog

- Move some functionality from NestUI into ConsoleLog

- Move the remainder into AbstractRunner

- Improve TestTranscript formatting

- Remove some unused methods
