+++ src/partest/scala/tools/partest/nest/AbstractRunner.scala
@@ -20,24 +20,17 @@ import java.util.concurrent.TimeUnit.NANOSECONDS
 
 class AbstractRunner(val config: RunnerSpec.Config, protected final val testSourcePath: String, val fileManager: FileManager) {
 
-  // lazy because colorEnabled is overridden in SBTRunner
-  // TODO: integrate NestUI into AbstractRunner
-  private lazy val nestUI: NestUI = new NestUI(
-    verbose = config.optVerbose,
-    debug = config.optDebug || propOrFalse("partest.debug"),
-    terse = config.optTerse,
-    colorEnabled = colorEnabled
-  )
-
   val javaCmdPath: String            = PartestDefaults.javaCmd
   val javacCmdPath: String           = PartestDefaults.javacCmd
   val scalacExtraArgs: Seq[String]   = Seq.empty
   val javaOpts: String               = PartestDefaults.javaOpts
   val scalacOpts: String             = PartestDefaults.scalacOpts
+  val debug: Boolean                 = config.optDebug || propOrFalse("partest.debug")
+  val verbose: Boolean               = config.optVerbose
+  val terse: Boolean                 = config.optTerse
 
   protected val printSummary         = true
   protected val partestCmd           = "test/partest"
-  protected val colorEnabled         = sys.props contains "partest.colors"
 
   private[this] var totalTests       = 0
   private[this] val passedTests      = mutable.ListBuffer[TestState]()
@@ -48,9 +41,70 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
   private[this] var expectedFailures = 0
   private[this] var onlyIndividualTests = false
 
+  val log = new ConsoleLog(sys.props contains "partest.colors")
+  import log._
 
-  import nestUI._
-  import nestUI.color._
+  private[this] val testNum = new java.util.concurrent.atomic.AtomicInteger(1)
+  @volatile private[this] var testNumberFmt = "%3d"
+  private[this] def testNumber = testNumberFmt format testNum.getAndIncrement()
+  def resetTestNumber(max: Int = -1): Unit = {
+    testNum set 1
+    val width = if (max > 0) max.toString.length else 3
+    testNumberFmt = s"%${width}d"
+  }
+
+  private[this] val realSysErr = System.err
+
+  def statusLine(state: TestState, durationMs: Long) = {
+    import state._
+    import TestState._
+    val colorizer = state match {
+      case _: Skip     => yellow
+      case _: Updated  => cyan
+      case s if s.isOk => green
+      case _           => red
+    }
+    val word = bold(colorizer(state.shortStatus))
+    def durationString = if (durationMs > PartestDefaults.printDurationThreshold) f"[duration ${(1.0 * durationMs) / 1000}%.2fs]" else ""
+    f"$word $testNumber - $testIdent%-40s$reasonString$durationString"
+  }
+
+  def reportTest(state: TestState, info: TestInfo, durationMs: Long, diffOnFail: Boolean, logOnFail: Boolean): Unit = {
+    if (terse && state.isOk) printDot()
+    else {
+      echo(statusLine(state, durationMs))
+      if (!state.isOk) {
+        def showLog() = if (info.logFile.canRead) {
+          echo(bold(cyan(s"##### Log file '${info.logFile}' from failed test #####\n")))
+          echo(info.logFile.fileContents)
+        }
+        if (diffOnFail) {
+          val differ = bold(red("% ")) + "diff "
+          val diffed = state.transcript find (_ startsWith differ)
+          diffed match {
+            case Some(diff) => echo(diff)
+            case None if !logOnFail && !verbose => showLog()
+            case _ => ()
+          }
+        }
+        if (logOnFail) showLog()
+      }
+    }
+  }
+
+  def usage(): Unit = {
+    println(RunnerSpec.programInfo.usage)
+    println(RunnerSpec.helpMsg)
+    sys.exit(1)
+  }
+
+  def verbose(msg: => String): Unit =
+    if (verbose) realSysErr.println(msg)
+
+  def showAllJVMInfo(): Unit = {
+    verbose(vmArgString)
+    verbose(allPropertiesString)
+  }
 
   private[this] def comment(s: String) = echo(magenta("# " + s))
 
@@ -89,7 +143,7 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
       val message   = passFail + elapsed
 
       if (failed0.nonEmpty) {
-        if (nestUI.verbose) {
+        if (verbose) {
           echo(bold(cyan("##### Transcripts from failed tests #####\n")))
           failed0 foreach { state =>
             comment(partestCmd + " " + state.testFile)
@@ -117,20 +171,20 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
     PathSettings.testSourcePath = testSourcePath
 
     if (config.optVersion) echo(versionMsg)
-    else if (config.optHelp) nestUI.usage()
+    else if (config.optHelp) usage()
     else {
       val (individualTests, invalid) = config.parsed.residualArgs map (p => Path(p)) partition denotesTestPath
       if (invalid.nonEmpty) {
-        if (nestUI.verbose)
+        if (verbose)
           invalid foreach (p => echoWarning(s"Discarding invalid test path " + p))
-        else if (!nestUI.terse)
+        else if (!terse)
           echoWarning(s"Discarding ${invalid.size} invalid test paths")
       }
 
       config.optTimeout foreach (x => setProp("partest.timeout", x))
 
-      if (!nestUI.terse)
-        nestUI.echo(banner)
+      if (!terse)
+        echo(banner)
 
       val grepExpr = config.optGrep getOrElse ""
 
@@ -227,7 +281,7 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
 
   def runTest(testFile: File): TestState = {
     val start = System.nanoTime()
-    val runner = new Runner(testFile, this, nestUI)
+    val runner = new Runner(testFile, this)
     var stopwatchDuration: Option[Long] = None
 
     // when option "--failed" is provided execute test only if log
@@ -242,7 +296,7 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
           case t: Throwable => throw new RuntimeException(s"Error running $testFile", t)
         }
       stopwatchDuration = Some(durationMs)
-      nestUI.reportTest(state, runner, durationMs, diffOnFail = config.optShowDiff || onlyIndividualTests , logOnFail = config.optShowLog || onlyIndividualTests)
+      reportTest(state, runner, durationMs, diffOnFail = config.optShowDiff || onlyIndividualTests , logOnFail = config.optShowLog || onlyIndividualTests)
       runner.cleanup()
       state
     }
@@ -252,7 +306,7 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
   }
 
   def runTestsForFiles(kindFiles: Array[File], kind: String): Array[TestState] = {
-    nestUI.resetTestNumber(kindFiles.size)
+    resetTestNumber(kindFiles.size)
 
     val pool              = Executors newFixedThreadPool PartestDefaults.numThreads
     val futures           = kindFiles map (f => pool submit callable(runTest(f.getAbsoluteFile)))
@@ -265,9 +319,9 @@ class AbstractRunner(val config: RunnerSpec.Config, protected final val testSour
       case Failure(e) =>
         e match {
           case TimeoutException(d)      =>
-            nestUI.warning("Thread pool timeout elapsed before all tests were complete!")
+            warning("Thread pool timeout elapsed before all tests were complete!")
           case ie: InterruptedException =>
-            nestUI.warning("Thread pool was interrupted")
+            warning("Thread pool was interrupted")
             ie.printStackTrace()
         }
         pool.shutdownNow()     // little point in continuing
